Refs #21736: Updates to support Fast DDS 3.x in SustainML-Framework
